+++ customxepr/gui/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
-from .manager_ui import ManagerApp
-from .customxepr_ui import CustomXeprGuiApp
+from .manager_window import ManagerApp
+from .customxepr_window import CustomXeprGuiApp
 from .splash import SplashScreen
