+++ customxepr/gui/manager_window_linux.ui
