+++ customxepr/gui/manager_window_macos.ui
